requirements.txt
# MCP Server Dependencies

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Configuration
python-dotenv>=1.0.0

# HTTP client (for fetch_webpage)
httpx>=0.25.0

# HTML to Markdown conversion
markdownify>=0.11.0
beautifulsoup4>=4.12.0

# Gmail integration (Phase 2)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.108.0

# Discord integration
discord.py>=2.3.0

# LinkedIn integration (unofficial API)
linkedin-api>=2.0.0
browser-cookie3>=0.19.0

